git-svn-id: svn://svn.code.sf.net/p/jsidplay2/code/trunk@2732 05cac4c1-a8e4-46c3-9d4b-80028926d4ef

+++ jsidplay2/src/main/java/ui/entities/whatssid/service/WhatsSidService.java
@@ -96,7 +96,7 @@ public class WhatsSidService implements FingerPrintingDataSource {
 		CriteriaBuilder cb = em.getCriteriaBuilder();
 		CriteriaQuery<MusicInfo> query = cb.createQuery(MusicInfo.class);
 		Root<MusicInfo> musicInfo = query.from(MusicInfo.class);
-		Path<Long> idMusicInfo = musicInfo.<Long>get(MusicInfo_.idMusicInfo);
+		Path<Integer> idMusicInfo = musicInfo.<Integer>get(MusicInfo_.idMusicInfo);
 		Predicate predicate = cb.equal(idMusicInfo, songNoBean.getSongNo());
 		query.where(predicate).select(musicInfo);
